Move TestDataContainerAlgebra to Utilities.py
