+++ src/xGadgetron/pGadgetron/tests/test_algebra.py
@@ -22,11 +22,11 @@ import os
 import shutil
 import unittest
 import sirf.Gadgetron as mr
-from sirf.Utilities import  examples_data_path
+from sirf.Utilities import  examples_data_path, TestDataContainerAlgebra
 
 
 
-class TestGadgetronAcquisitionDataAlgebra(unittest.TestCase):
+class TestGadgetronAcquisitionDataAlgebra(unittest.TestCase, TestDataContainerAlgebra):
 
     def setUp(self):
         os.chdir(examples_data_path('MR'))
@@ -44,184 +44,9 @@ class TestGadgetronAcquisitionDataAlgebra(unittest.TestCase):
         pass
 
 
-    def test_divide_scalar(self):
-        os.chdir(self.cwd)
-        image1 = self.image1
-        image2 = self.image2
 
-        image1.fill(1.)
-        image2.fill(2.)
-        
-        tmp = image1/1.
-        numpy.testing.assert_array_equal(image1.as_array(), tmp.as_array())
-    
-        tmp1 = image1.divide(1.)
-        numpy.testing.assert_array_equal(tmp.as_array(), tmp1.as_array())
-        
-        image1.divide(1., out=image2)
-        numpy.testing.assert_array_equal(tmp.as_array(), image2.as_array())
-        
-    def test_divide_datacontainer(self):
-        os.chdir(self.cwd)
-        
-        image1 = self.image1
-        image2 = self.image2
-        
 
-        # add 1 because the data contains zeros and divide is not going to be happy
-        image1+=1.
-        image2+=1.
-        tmp = image1/image2
-
-        numpy.testing.assert_array_almost_equal(
-            numpy.ones(image1.shape, dtype=numpy.float32), tmp.as_array()
-            )
-    
-        tmp1 = image1.divide(image2)
-        numpy.testing.assert_array_almost_equal(
-            numpy.ones(image1.shape, dtype=numpy.float32), tmp1.as_array()
-            )
-        
-        tmp1.fill(2.)
-        image1.divide(image2, out=tmp1)
-        
-        numpy.testing.assert_array_almost_equal(
-            numpy.ones(image1.shape, dtype=numpy.float32), tmp1.as_array()
-            )
-        
-
-    def test_multiply_scalar(self):
-        os.chdir(self.cwd)
-        image1 = self.image1
-        image2 = self.image2
-        
-
-        image2.fill(2.)
-        
-        tmp = image1 * 1.
-        numpy.testing.assert_array_equal(image1.as_array(), tmp.as_array())
-    
-        tmp1 = image1.multiply(1.)
-        numpy.testing.assert_array_equal(tmp.as_array(), tmp1.as_array())
-        
-        image1.multiply(1., out=image2)
-        numpy.testing.assert_array_equal(tmp.as_array(), image2.as_array())
-        
-    def test_multiply_datacontainer(self):
-        os.chdir(self.cwd)
-        image1 = self.image1
-        image2 = self.image2
-        
-
-        image2.fill(1.)
-        tmp = image1 * image2
-
-        numpy.testing.assert_array_almost_equal(
-            image1.as_array(), tmp.as_array()
-            )
-    
-        tmp1 = image1.multiply(image2)
-        numpy.testing.assert_array_almost_equal(
-            image1.as_array(), tmp1.as_array()
-            )
-        
-        tmp1.fill(2.)
-        image1.multiply(image2, out=tmp1)
-        
-        numpy.testing.assert_array_almost_equal(
-            image1.as_array(), tmp1.as_array()
-            )
-        
-    def test_add_scalar(self):
-        os.chdir(self.cwd)
-        image1 = self.image1
-        image2 = self.image2
-        
-        image1.fill(0)
-        image2.fill(1)
-        
-        tmp = image1 + 1.
-        numpy.testing.assert_array_equal(image2.as_array(), tmp.as_array())
-    
-        tmp1 = image1.add(1.)
-        numpy.testing.assert_array_equal(tmp.as_array(), tmp1.as_array())
-        
-        tmp1.fill(0)
-        image1.add(1., out=tmp1)
-        numpy.testing.assert_array_equal(tmp1.as_array(), image2.as_array())
-    
-    def test_add_datacontainer(self):
-        os.chdir(self.cwd)
-        image1 = self.image1
-        image2 = self.image2
-        
-
-        image1.fill(0.)
-        image2.fill(1.)
-        tmp = image1 + image2
-
-        numpy.testing.assert_array_almost_equal(
-            numpy.ones(image1.shape, dtype=numpy.float32), tmp.as_array()
-            )
-    
-        tmp1 = image1.add(image2)
-        
-        numpy.testing.assert_array_almost_equal(
-            numpy.ones(image1.shape, dtype=numpy.float32), tmp1.as_array()
-            )
-        
-        tmp1.fill(2.)
-        image1.add(image2, out=tmp1)
-        
-        numpy.testing.assert_array_almost_equal(
-            numpy.ones(image1.shape, dtype=numpy.float32), tmp1.as_array()
-            )
-        
-    
-    def test_subtract_scalar(self):
-        os.chdir(self.cwd)
-        image1 = self.image1
-        image2 = self.image2
-        
-
-        image1.fill(2)
-        image2.fill(1)
-        
-        tmp = image1 - 1.
-        numpy.testing.assert_array_equal(image2.as_array(), tmp.as_array())
-    
-        tmp1 = image1.subtract(1.)
-        numpy.testing.assert_array_equal(tmp.as_array(), tmp1.as_array())
-        
-        tmp1.fill(0)
-        image1.subtract(1., out=tmp1)
-        numpy.testing.assert_array_equal(tmp1.as_array(), image2.as_array())
-    
-    def test_subtract_datacontainer(self):
-        os.chdir(self.cwd)
-        image1 = self.image1
-        image2 = self.image2
-        
-        tmp = image1 - image2
-
-        numpy.testing.assert_array_almost_equal(
-            numpy.zeros(image1.shape, dtype=numpy.float32), tmp.as_array()
-            )
-    
-        tmp1 = image1.subtract(image2)
-        
-        numpy.testing.assert_array_almost_equal(
-            numpy.zeros(image1.shape, dtype=numpy.float32), tmp1.as_array()
-            )
-        
-        tmp1.fill(2.)
-        image1.subtract(image2, out=tmp1)
-        
-        numpy.testing.assert_array_almost_equal(
-            numpy.zeros(image1.shape, dtype=numpy.float32), tmp1.as_array()
-            )
-
-class TestGadgetronImageDataAlgebra(TestGadgetronAcquisitionDataAlgebra):
+class TestGadgetronImageDataAlgebra(unittest.TestCase, TestDataContainerAlgebra):
 
     def setUp(self):
         os.chdir(examples_data_path('MR'))
@@ -236,3 +61,8 @@ class TestGadgetronImageDataAlgebra(TestGadgetronAcquisitionDataAlgebra):
         self.image1 = image_data
         self.image2 = image_data * 1
         self.cwd = os.getcwd()
+
+    def tearDown(self):
+        #shutil.rmtree(self.cwd)
+        pass
+
